Split tokenizer into separate file, plus fix no_std feature.
